backend/pyproject.toml
[project]
name = "amorae-backend"
version = "1.0.0"
description = "Amorae AI Companion Backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-dotenv>=1.0.0",
    "firebase-admin>=6.4.0",
    "openai>=1.10.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "asyncpg>=0.29.0",
    "sqlalchemy>=2.0.25",
    "pgvector>=0.2.4",
    "redis>=5.0.0",
    "google-cloud-firestore>=2.14.0",
    "google-cloud-storage>=2.14.0",
    "google-cloud-tasks>=2.15.0",
    "sse-starlette>=1.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
